requirements.txt
asgiref==3.8.1
Django==5.1.1
django-autoslug==1.9.9
djangorestframework==3.15.2
drf-yasg==1.21.7
gunicorn==20.1.0
inflection==0.5.1
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
setuptools==75.1.0
sqlparse==0.5.1
uritemplate==4.1.1
whitenoise==6.4.0
